refactor(Timer.cs): Simplify timer class
There were some weird things about this class while I was using it in
the game I'm making with this library, so I decided to refactor it and
then add unit tests for it so I don't have to worry about it ever again.

The main and only change is that when you create the class you don't
have to immediately set it to be full, it will now just start at max.