refactor(Timer.cs): Another general timer refactor
Since there are so many possible deltas that you could use with the
timer everything got consolidated to the Timer.Tick() method, which
allows you to use whatever delta you want.

In addition, the class could not be serialized for easier configuration,
this has been fixed.

A missing feature was "auto restart" which ideally would restart the timer
automatically as it runs out. This feature has been added.